feat: removed old text input comp file
